requirements.txt
autopep8==1.5.4
flake8==3.8.3
gitdb==4.0.5
GitPython==3.1.8
mccabe==0.6.1
pycodestyle==2.6.0
pyflakes==2.2.0
smmap==3.0.4
toml==0.10.1
